# src/AppForSEII2526.API/Controllers/HerramientaController.cs
using AppForSEII2526.API.DTOs;
using AppForSEII2526.API.DTOs;
using AppForSEII2526.API.Models;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using SQLitePCL;
using System.Linq;
using static System.Runtime.InteropServices.JavaScript.JSType;

namespace AppForSEII2526.API.Controllers

    {



        [Route("api/[controller]")]
        [ApiController]
        public class HerramientaController : ControllerBase
        {
            private readonly ApplicationDbContext _context;
            private readonly ILogger<HerramientaController> logger;

            public HerramientaController(ApplicationDbContext context, ILogger<HerramientaController> logger)
            {
                _context = context;
                this.logger = logger;
               
            }


            [HttpGet]
            [Route("[action]")]
            [ProducesResponseType(typeof(IList<HerramienParaAlquilarDTO>), (int)HttpStatusCode.OK)]
            public async Task<ActionResult> GetHerramientasForRenting(string? nombre, string? material)
            {
                DateTime pasMñn= DateTime.Now.AddDays(2), semSig= DateTime.Now.AddDays(7);
   
            IList<HerramienParaAlquilarDTO> selectherr = await _context.Herramienta
                .Include(alq => alq.AlquilarItems).ThenInclude(al => al.alquiler) 
                    .Where (alq => (nombre==null || alq.nombre.Contains(nombre)) 
                    && (material==null || alq.material.Contains(material))
                    && (alq.AlquilarItems.All(a => a.alquiler.fechaInicio.Date > semSig.Date 
                        || a.alquiler.fechaFin.Date < pasMñn.Date)))
                    .OrderBy(alq => alq.Id)
                    .Select(alq => new HerramienParaAlquilarDTO(alq.Id, alq.material, alq.nombre, alq.precio, alq.fabricante.nombre))
                    .ToListAsync();

            if (selectherr.Count == 0)
            {
                throw new Exception("No se encontraron herramientas disponibles para alquilar con los filtros proporcionados.");
                
            }
                return Ok(selectherr);
            }

        [HttpGet]
        [Route("[action]")]
        [ProducesResponseType(typeof(IList<HerramienParaComprarDTO>), (int)HttpStatusCode.OK)]
        [ProducesResponseType((int)HttpStatusCode.NotFound)]
        public async Task<IActionResult> GetHerramienParaComprar(string? material, float? precio, string? nombreHerramienta)
        {
            if (_context.Herramienta == null)
            {
                logger.LogWarning("No se encontraron herramientas en la base de datos.");
                return NotFound();
            }
            logger.LogInformation("Iniciando GetHerramienParaComprar");
            var herramientas = await _context.Herramienta
                .Include(h => h.fabricante)
                .Where(h => (material == null || h.material.ToLower().Contains(material.ToLower())) &&
                            (precio == null || h.precio == precio) && (nombreHerramienta == null || h.nombre.ToLower().Contains(nombreHerramienta.ToLower())))
                .Select(h => new HerramienParaComprarDTO(
                    h.Id,
                    h.material,
                    h.nombre,
                    h.precio,
                    h.fabricante != null ? h.fabricante.nombre : string.Empty
                ))
                .ToListAsync();
            logger.LogInformation("Finalizando GetHerramienParaComprar");
            return Ok(herramientas);
        }
       
        [HttpGet]
        [Route("[action]")]
        [ProducesResponseType(typeof(IList<HerramientaparaOfertaDTO>), (int)HttpStatusCode.OK)]
        public async Task<IActionResult> GetHerramientasParaOferta(string? fabricante, float? precio)
        {
            IList<HerramientaparaOfertaDTO> selectherr = await _context.Herramienta
                .Include(Ofer=> Ofer.OfertaItem).ThenInclude(Ofe => Ofe.oferta)
                    .Where(Ofer => (fabricante == null || Ofer.fabricante.nombre.Contains(fabricante))
                    && (precio == null || Ofer.precio<=precio))
                    .OrderBy(Ofer => Ofer.nombre)
                    .Select(Ofer=> new HerramientaparaOfertaDTO(Ofer.Id, Ofer.material, Ofer.nombre, Ofer.precio, Ofer.fabricante.nombre))
                    .ToListAsync();
            return Ok(selectherr);
        }


        [HttpGet]
        [Route("[action]")]
        [ProducesResponseType(typeof(IList<HerramientaParaRepararDTO>), (int)HttpStatusCode.OK)]
        public async Task<ActionResult> GetHerramientasParaReparar(string? filtroNombre, int? filtroTiempoReparacion)
        {

            var herramientas = await _context.Herramienta
                .Include(h => h.Fabricante)
                .Where(h => (h.nombre.Contains(filtroNombre) || filtroNombre == null)
                    && (h.tiempoReparacion <= filtroTiempoReparacion || filtroTiempoReparacion == null))
                .OrderBy(h => h.tiempoReparacion)
                .Select(h => new HerramientaParaRepararDTO(
                    h.Id,
                    h.nombre,
                    h.material,
                    h.precio,
                    h.tiempoReparacion,
                    h.Fabricante.nombre
                )).ToListAsync();
            return Ok(herramientas);
        }
    }
}