# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Herramienta/GetHerramientasForRenting": {
      "get": {
        "tags": [
          "Herramienta"
        ],
        "operationId": "GetHerramientasForRenting",
        "parameters": [
          {
            "name": "nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "material",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramienParaAlquilarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramienParaAlquilarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramienParaAlquilarDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Herramienta/GetHerramienParaComprar": {
      "get": {
        "tags": [
          "Herramienta"
        ],
        "operationId": "GetHerramienParaComprar",
        "parameters": [
          {
            "name": "material",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "precio",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "nombreHerramienta",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramienParaComprarDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramienParaComprarDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramienParaComprarDTO"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Herramienta/GetHerramientasParaOferta": {
      "get": {
        "tags": [
          "Herramienta"
        ],
        "operationId": "GetHerramientasParaOferta",
        "parameters": [
          {
            "name": "fabricante",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "precio",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaparaOfertaDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaparaOfertaDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaparaOfertaDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Herramienta/GetHerramientasParaReparar": {
      "get": {
        "tags": [
          "Herramienta"
        ],
        "operationId": "GetHerramientasParaReparar",
        "parameters": [
          {
            "name": "filtroNombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filtroTiempoReparacion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaParaRepararDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaParaRepararDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HerramientaParaRepararDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HerramienParaAlquilarDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "material": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "precio": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HerramienParaComprarDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "material": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "maxLength": 50,
            "minLength": 10,
            "type": "string",
            "nullable": true
          },
          "precio": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HerramientaParaRepararDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "material": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "precio": {
            "minimum": 0.05,
            "type": "number",
            "format": "currency"
          },
          "tiempoReparacion": {
            "type": "number",
            "format": "float"
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HerramientaparaOfertaDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}